+++ src/util/TypeTraits.h
@@ -215,6 +215,10 @@ template <typename... Ts>
 requires(sizeof...(Ts) > 0)
 using First = typename detail::FirstWrapper<Ts...>::type;
 
+/// Concept for `std::is_invocable_r_v`.
+template <typename Func, typename R, typename... ArgTypes>
+concept InvocableWithReturnType = std::is_invocable_r_v<R, Func, ArgTypes...>;
+
 /*
 The following concepts are similar to `std::is_invocable_r_v` with the following
 difference: `std::is_invocable_r_v` only checks that the return type of a
