+++ test/OperationTest.cpp
@@ -9,10 +9,10 @@
 #include "engine/NeutralElementOperation.h"
 #include "engine/ValuesForTesting.h"
 #include "util/IdTableHelpers.h"
+#include "util/OperationTestHelpers.h"
 
 using namespace ad_utility::testing;
 using namespace ::testing;
-using namespace std::chrono_literals;
 using ad_utility::CancellationException;
 using ad_utility::CancellationHandle;
 using ad_utility::CancellationState;
@@ -138,89 +138,6 @@ TEST_F(OperationTestFixture, verifyCachePreventsInProgressState) {
           ParsedAsJson(HasKeyMatching("status", Eq("not started"))),
           ParsedAsJson(HasKeyMatching("status", Eq("fully materialized")))));
 }
-// _____________________________________________________________________________
-
-class StallForeverOperation : public Operation {
-  std::vector<QueryExecutionTree*> getChildren() override { return {}; }
-  string asStringImpl([[maybe_unused]] size_t) const override {
-    return "StallForEverOperation";
-  }
-  string getDescriptor() const override {
-    return "StallForEverOperationDescriptor";
-  }
-  size_t getResultWidth() const override { return 0; }
-  size_t getCostEstimate() override { return 0; }
-  uint64_t getSizeEstimateBeforeLimit() override { return 0; }
-  float getMultiplicity([[maybe_unused]] size_t) override { return 0; }
-  bool knownEmptyResult() override { return false; }
-  vector<ColumnIndex> resultSortedOn() const override { return {}; }
-  VariableToColumnMap computeVariableToColumnMap() const override { return {}; }
-
- public:
-  using Operation::Operation;
-  // Do-nothing operation that runs for 100ms without computing anything, but
-  // which can be cancelled.
-  ResultTable computeResult() override {
-    auto end = std::chrono::steady_clock::now() + 100ms;
-    while (std::chrono::steady_clock::now() < end) {
-      checkCancellation();
-    }
-    throw std::runtime_error{"Loop was not interrupted for 100ms, aborting"};
-  }
-
-  // Provide public view of remainingTime for tests
-  std::chrono::milliseconds publicRemainingTime() const {
-    return remainingTime();
-  }
-};
-// _____________________________________________________________________________
-
-// Dummy parent to test recursive application of a function
-class ShallowParentOperation : public Operation {
-  std::shared_ptr<QueryExecutionTree> child_;
-
-  explicit ShallowParentOperation(std::shared_ptr<QueryExecutionTree> child)
-      : child_{std::move(child)} {}
-  string asStringImpl([[maybe_unused]] size_t) const override {
-    return "ParentOperation";
-  }
-  string getDescriptor() const override { return "ParentOperationDescriptor"; }
-  size_t getResultWidth() const override { return 0; }
-  size_t getCostEstimate() override { return 0; }
-  uint64_t getSizeEstimateBeforeLimit() override { return 0; }
-  float getMultiplicity([[maybe_unused]] size_t) override { return 0; }
-  bool knownEmptyResult() override { return false; }
-  vector<ColumnIndex> resultSortedOn() const override { return {}; }
-  VariableToColumnMap computeVariableToColumnMap() const override { return {}; }
-
- public:
-  template <typename ChildOperation, typename... Args>
-  static ShallowParentOperation of(QueryExecutionContext* qec, Args&&... args) {
-    return ShallowParentOperation{
-        ad_utility::makeExecutionTree<ChildOperation>(qec, args...)};
-  }
-
-  std::vector<QueryExecutionTree*> getChildren() override {
-    return {child_.get()};
-  }
-
-  ResultTable computeResult() override {
-    auto childResult = child_->getResult();
-    return {childResult->idTable().clone(), resultSortedOn(),
-            childResult->getSharedLocalVocab()};
-  }
-
-  // Provide public view of remainingTime for tests
-  std::chrono::milliseconds publicRemainingTime() const {
-    return remainingTime();
-  }
-};
-
-template <>
-void QueryExecutionTree::setOperation<StallForeverOperation>(
-    std::shared_ptr<StallForeverOperation> operation) {
-  _rootOperation = std::move(operation);
-}
 
 // _____________________________________________________________________________
 
@@ -231,35 +148,32 @@ TEST(OperationTest, verifyExceptionIsThrownOnCancellation) {
       ShallowParentOperation::of<StallForeverOperation>(qec);
   operation.recursivelySetCancellationHandle(handle);
 
-  std::thread thread{[&]() {
+  ad_utility::JThread thread{[&]() {
     std::this_thread::sleep_for(5ms);
     handle->cancel(CancellationState::TIMEOUT);
   }};
-  EXPECT_THROW(
-      {
-        try {
-          operation.computeResult();
-        } catch (const ad_utility::AbortException& exception) {
-          EXPECT_THAT(exception.what(),
-                      ::testing::HasSubstr("Cancelled due to timeout"));
-          throw;
-        }
-      },
+  AD_EXPECT_THROW_WITH_MESSAGE_AND_TYPE(
+      operation.computeResult(),
+      ::testing::HasSubstr("Cancelled due to timeout"),
       ad_utility::AbortException);
-  thread.join();
 }
 
 // _____________________________________________________________________________
 
 TEST(OperationTest, verifyRemainingTimeDoesCountDown) {
+  constexpr auto timeout = 5ms;
   auto qec = getQec();
   ShallowParentOperation operation =
       ShallowParentOperation::of<StallForeverOperation>(qec);
-  operation.recursivelySetTimeConstraint(1ms);
-  std::this_thread::sleep_for(1ms);
-  // Verify time is up for parent and child
-  EXPECT_EQ(operation.publicRemainingTime(), 0ms);
+  operation.recursivelySetTimeConstraint(timeout);
+
   auto childOperation = std::dynamic_pointer_cast<StallForeverOperation>(
       operation.getChildren().at(0)->getRootOperation());
+
+  EXPECT_GT(operation.publicRemainingTime(), 0ms);
+  EXPECT_GT(childOperation->publicRemainingTime(), 0ms);
+  std::this_thread::sleep_for(timeout);
+  // Verify time is up for parent and child
+  EXPECT_EQ(operation.publicRemainingTime(), 0ms);
   EXPECT_EQ(childOperation->publicRemainingTime(), 0ms);
 }
